% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panic.R
\name{dontPanic}
\alias{dontPanic}
\title{Don't Panic}
\usage{
dontPanic(mins = 60, banner = TRUE)
}
\arguments{
\item{mins}{For how many minutes should the alarm continue to sound? Defaults to 60.}

\item{banner}{Show a "PANIC" banner message in ASCII art? Defaults to TRUE.}
}
\value{

}
\description{
Displays a green ASCII banner.
}
