% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dontPanic.R
\name{panic}
\alias{panic}
\title{Panic}
\usage{
panic(mins = 60, banner = TRUE)
}
\arguments{
\item{mins}{For how many minutes should the alarm continue to sound? Defaults to 60.}

\item{banner}{Show a "PANIC" banner message in ASCII art? Defaults to TRUE.}
}
\value{

}
\description{
Sounds an audio alarm, and optionally displays a red ASCII banner.
}
