% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panicWhen.R
\name{panicWhen}
\alias{panicWhen}
\title{Panic When}
\usage{
panicWhen(
  panic_condtion,
  mins = 60,
  panic_banner = TRUE,
  dontpanic_banner = TRUE,
  print_time = TRUE
)
}
\arguments{
\item{mins}{In case of panic, for how many minutes should the alarm continue to sound? Defaults to 60.}

\item{panic_banner}{In panic condition, show a "PANIC" banner message in ASCII art? Defaults to TRUE.}

\item{dontpanic_banner}{In don't panic condition, show a "DON'T PANIC" banner message in ASCII art? Defaults to TRUE.}

\item{print_time}{Print the system time before checking the panic condition (useful for verifying panic check is active). Defaults to TRUE.}

\item{panic_condition}{Logical that is TRUE if a panic should occur, FALSE when no panic should occur.}
}
\value{

}
\description{
When the condition is met, sound an audio alarm, and optionally display a red "PANIC" banner.  When the condition is not met, optionally display a "DON'T PANIC" banner.
}
