% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redAlert.R
\name{redAlert}
\alias{redAlert}
\title{Red Alert}
\usage{
redAlert(mins = 60, banner = TRUE)
}
\arguments{
\item{mins}{For how many minutes should the alarm continue to sound? Defaults to 60.}

\item{banner}{Show a "Red Alert" banner message in ASCII art? Defaults to TRUE.}
}
\value{

}
\description{
Sounds the Red Alert sound effect from Star Trek: The Next Generation, and optionally displays a red ASCII banner.
}
